Added matrix multiplication

+++ modules/Parcs.Modules.ProofOfWork/ModuleOptions.cs
@@ -6,9 +6,9 @@
 
         public string Prompt { get; set; } = "Hello world!";
 
-        public int NonceBatchSize { get; set; } = 1000;
+        public long NonceBatchSize { get; set; } = 1000;
 
-        public int MaximumNonce { get; set; } = 3000000;
+        public long MaximumNonce { get; set; } = 3000000;
 
         public string OutputFilename { get; set; } = "Output.txt";
     }
