+++ modules/Parcs.Modules.ProofOfWork/ModuleOutput.cs
@@ -6,7 +6,7 @@
 
         public bool Found { get; set; }
 
-        public int? ResultNonce { get; set; }
+        public long? ResultNonce { get; set; }
 
         public string ResultHash { get; set; }
     }
